Cell tile bugfix, implemented tileset colorization
When placing a tile in Cell Editing Mode, the incorrect tile would be placed. Tile indices in the cell file are always 0-indexed, but the first texture generated in the program may not be the tiles themselves (the tile color choosing buttons are the first to be created). This offset discrepancy was fixed by calculating the tile texture offset (or palette texture offset) at placement time and then subtracting that offset from the current tile texture. This offset subtraction operation was left in the cell tile placement code (as opposed to subtracting it when `current_tile` is assigned) as a reminder that the tile offsets in the cell file must be corrected.

Also made the tileset and the current tile indicator colorized if the `Use Palette` checkbox is enabled.