Added more output formats, fixed scrolling issue
So I just recently learned that RGBDS is slightly stupid (but it's trying its best!) and doesn't understand the `069h` syntax, so I implemented the ability to choose between IntSys (069h), Rednex ($69), and Binary (0x69) output formats when saving files. 

Also, fixed an issue with placing tiles and cells when scrolling. (e.g. when working on a large canvas)

Bumping an entire minor version due to supporting the superior "modern tooling" of RGBDS. ;)